LayoutTests/fast/frames/adopt-object-into-itself-expected.txt

ALERT: HierarchyRequestError (DOM Exception 3): The operation would yield an incorrect node tree.
